Setup before and after render for manager
The manager now can call before and after render to do setup and tear
down before rendering each entity. This has to be done because SDL
needs to do things every loop like clearing the scene.

Another thing was noticied: entity needs a reference to the renderer
because the components are expecting to be able to call it.