Added death test for exception condition
This was to test that in the case hwere a component isn't found,
the program exists.

In reality, it seems better to throw an exception, or be able to
control what happens in different environments. So on prod it would
exit with an user friendly error message but on dev it would show
a stack trace. Not sure how to do this with google death tests.