Add interface for entity
Start using it in manager.

I'm doing this for different reasons. The main reason is for testing,
it's much easier to test if theres an inteface. I think theres a lib
out there that could mock non-virtual methods on an object (GoogleMock)
can't, but I'd rather do it the more correct way by using an interface.

I also want to learn more about dependency injected code.

TODO:
The box, v and pos members on Entity should be moved to private
and have accessors. I realized this more now with this interface
as it prohibits using virtual methods for all parts of the object.