Getting entities of type
This function was written to allow an interface for easily getting
all entities that have a certain component. It will likely be
used internally to get all collision entities. The internal
implementation can be changed to a std::map with string to
std::vector of entities, which would be faster but more work to
maintain.