Created system class
Decided to only have an update function because I looked at
potential systems and very few needed rendering as a separate
step at this point. I may subclass system to rendersystem for
systems that require rendering. Then the manager add system function
can be overloaded to accept either type.

Used abstract class like this so getType can be inherited. Techincally
I think update could be pure virtual, but don't want to do it at
this point so I dont' have to use it around as a pointer all the
time.