Clock re-implementation
Not testing the starting and stopping functionality because this
application is single threaded right now, so control is never
returned to the caller. I think that might be the more correct
way as the app is going to stop the clock by the update functions
that are running on the clock. But that means this code may
not be testable.

The ticking seems to be working correctly.