Added test movement system and component
The tests for this are in one file, movement. This is because they
are closely related. The moveable component gets velocity now as
the component shoudl keep whatever data is necessary for  it to
render and update.

Prefixed the system vs the component with s vs c. This helps identify
what's what. Also named components with an adjective and systems
with a noun.

The Makefile here is a complete disaster. The whole make process
needs to be rewritten so it can use *.o:*.cpp and works from any
directory and for both tests.