Template relevant parts of the entity class
Components are based from teh same component type, but they add on
additional methods to be able to do their things, like getVelocity. This
means that using the base Component breaks when a system attempts to use
a component because the base doesn't have all those methods.

This attempts to fix this using templates and a static cast when you
get a component to use. This means that the entity interface has to be
removed and the manager class has to use a template for entity, to sub
in/out the entity mock.

It may also be possible to use a static_cast in the system, I might try
that next.

Question: is it possible to do this without a cast?