Got movement working
Changed makefile to ensure it re-compiles gaming.a every time.

Added a move method onto entity. I did this rather then regular
setter because I feel that move is a more realistic interface
to the entity class. When working with position, the most common
operation will be to move it with a velocity. A get accessor
should also be added.

Movement code all updated with new Entity class.