Added boundable component and tests
So now the world bounds will be kept in this component rather then
a random const like before. This value might have to be passed
around a little because mutliple places will need it.

Designed getBounds to return a const reference because, 1. its
fastest and 2 const works on a ref this time because bounds shouldn't
be changed (probably not), as it could break the game.