Rectangle component
Simply holds the box for the rectangle.

I decided the public interface should not set and get a 2d vector this
time because a 2d vector has x, and y members, which sounds a little
incorrect for a rectangle that should be width and height. I also felt
that these values would change individually and not at the same time
so this interface may be more appropriate.