Add color struct
This will add a layer of separation from SDL colors. This is done so
if the library implementation changes to something other then SDL, the
colors will not have to change.

One potential problem I see with this is that alpha is an uint8 rather
then a float, which may not be the most common implementation.