Painted component
I'm not 100% sure on the interface but do think it'd be useful to create
the component by just passing in numbers for rgba. It also might be use-
ful if you can pass in a color. Could end up making set/get rgba methods
as well, not sure if it will be easier.