Keyboardable component
The purpose of this components is to provide a key mapping to a
different system or component. So you can set constants ints for
certain actions, pass a keymap (which could include alias keys) to
keyboardable, and the keyboard_event system will use the keymap to
emit the correct events.

Kept the interface to what I thought was needed to use: setting an
initial map that should be const, getting all the keys in use and
getting the action.

TODO change getMovement to getAction.