Thrust
Thrust works by binding the entity to a the user movement const event
and manipulating the moveable component velocity vector directly.

There are some problems with this, like if the entity doesn't have a
moveable component. Another way would be to have thrust emit an event
for thrust that mvoeable has to listen to. Will determine after use
if this would be a lot better.

Something is wrong with Vector 2d *= operator as it's not working to
assign the factor right now. Will figure out later.