Add box2d body to moveable
I added aliases in an alias class so that my code stays clear of
box2d references. This means this could be replaced with something
else. Not convinced of this.

A pointer will be kept in moveable component that gets initialized
with the body def, which is created with initial values put into
constructor.