Create body in Moveable with world component
This allows using a component that keeps the world to create the body
on Moveable.

I tried this solution as an alternative to putting the world in Manager
and padding the world to init and onAddEntity because it limits the
amount of things that have to know about world, including Manager,
which shouldn't need to know about Box2d at all. The downside is
creating a full physics object means you always need to have universal,
which isn't that intuitive. This just made me realize that maybe it
should be called Physical as it relates directly to box2d, not world.

Fixes #16.