Add fields to moveable for b2fixture (-shape)
Decided to fix #20 by not having a separate physics component to
have to coordinate with, but put all of it in moveable (which should
be named physical). This means we don't ahve to coordinate creating
the fixture after the body was created.

This is not great design to have such a massive component, it should
eventually be fixed.

Fixes #20