Add fixture to moveable init
This is a mess, was done for fastest way to the end. The code has to
pass a shape to Moveable init so that it can add the shape to the
fixture. Does this in the movement system to keep in sync with how
the moveable body is created.

Problems:
- No way to apply one fixture to multiple bodys or multiple fixtures
to one body.
- Doesn't work for other shape types unless Moveable::init is templated
to accept different b2 shapes.
- Increases b2 dependencies on Moveable.