Start to have main shape component handle shapes
This attempts to have one shaped component handle getting the
b2shape and shape type for any corresponding shape (rectangular,etc).

This works by keeping a copy of rectangular or whatever to shaped,
which I did because it is an easy way to allow to force the constraint
of one shape per entity or shape component. This might have to
further worked with