Complete shape taking care of things
Using abstract base shape class to serve as base for which Shaped
component gives out pointer to. So now Shaped can give anything a
pointer to the main shape component which can be used to get the b2Shape
object which can be any type of b2 shape. This allows system that
need access to the shape, but shouldn't be aware of implementation
details, such as the differences between a rectangle and a polygon, to
just get the shape and past it along. Now any system can work with
different shapes

Fixes #32.