Add rendered component with texture
Moving texture from paint. This allows one central place where
various components needing the texture can access it and manipulate
it.

It was moved from painted because a texture would be required if
someting was an image or a sprite, as opposed to just a plain color,
which painted is supposed to do.