WIP Rectangle renderer setting up texture
This idea woud be to init the rendered component with a surface
which it can set and then create a texture if it wants.

THis won't work because you need a SDL renderer to create a texture
from a surface, so it might be better to have all the logic in
rectangle renderer and set the data in the rendered component. This
logic and be later moved to a different components if it needs to be
shared, such as a paint system to setup the texture for color vs
a img system to set up the texture for spriting.