Create texture for rendering
Creating a texture from a surface or at all requires an actual sdl
renderer, so I had no choice but move this into the SDL_Renderer and
renderer interface. I wanted to keep the IRenderer completely free of
any sdl specific code, so that part should be rewritten somehow. Maybe
the solution would be to use some sort of type alias like for vector.

Now the rectangle renderer system will create the surface and create
the texture by calling it's renderer. This was done here as it has
access to the sdl renderer, which is a dependency for some of this. It
is also already a little sdl specific, but it would be good if any SDL
code could be taken out.

The rendered component has one init method to setup both the texture and
surface together. I did this to apply a constraint that both are needed
but don't really like the design that much, but will wait to change for
when I render different shapes, as they'll bring new things to consider.