Render angle with texture
This uses the texture to use a different SDL render method that
allows to render angles.

Still have to work with the values for position and size from box2d
to sdl, as things aren't lining up correctly. Also, things aren't
rotating as much as I thought they would.

Fixes #35.