New event for user rotation
A specific event for when a player wants to rotate their ship.

I created it's own event class rather then add onto EvUserMovement
because movement takes a Vector2d as it's direction, which wouldn't
be able to express a rotation movement as opposed to a regular
movement. It feels a little hacky to modify user movement to also work
with rotation, so might make more sense to have separate events.

Using a float for the direction input as it will transfer nicely with
the float required to set angular velocity or apply a force in box2d.
Although might make more sense to use a smaller uint.

Fixes #48.