Finish keyboard events with rotation
I'm not convinced of this design. It requires creating two
KeyboardEvents systems of different template types and then using
typeid to check which one is use and not emit if it's not the
correct one. This allows to use one keymap with one component,
but I don't think it's great practice to use typeid.

Fixes #49.