Directional player controls
Changes the controls so they are relative to the current ship
angle.

Up and down controls are easily done by negating the x and y of the
force, so changing  the EvUserMovement vectors to both negative.

The right and left is a more complex of a calculation so requires
an if statement to set the correct angle based on an equation. Have
to figure out a better way to check for right or left that wouldn't
break if they're reset in main.

Fixes #54