Added impact movement
An impact event to represent when a certain impact has happened. It
follows box2d conventions of having multiple implulses that you can
combine to one float for the total impulse, which will be used to
calculate the amount of force, and then the amount of damage.

The Box2d listener will be calling this.

Fixes #63.