Work on box2d collision events
Takes box2d collision events and emits an impact if it's a damageable
body.

Uses a basic class in main for the listener to pass to box2d because
I haven't decided how this should be implemented yet and this was
easiest.

Damageable has a factor that gets applied to the damage, to match up
the health number with the impulse values that get returned by box2d.
Only the first normal impulse from box2d gets applied. The totalImpulses
isn't working correctly right now and I'm not sure what the different
ones represent. The factors will likely have to be modified when the
camera is set as scale will be different.

Fixes #64.