Fixed sdl ttf
Can now render text on multiple lines.

The problem with SDL_ttf was I had SDL1 ttf installed, so I was
including the wrong version. The include has to point to SDl2_ttf.
I also mistakenly was calling write text from update when it should
be called from render.

This method is very bad for performance, as it has to create a new
texture from a surface every time. Will have to investigate ways
to make this faster later.