Notification component to hold data
This will allow having different text outputs for two differnet
players in two different colors.

There will be a notify system that will use this component to print
the text to the screen in the respective place and color.

This system was designed to allow for more functionality in the future,
such as lines having different colors and lines having different
characters. It also allows a notify system to only create a new
texture when values have changed, allowing a more efficient system.