Fixed bounds with shield
Problem: the shield joint to the ship was breaking when crossing a world
boundary and being wrapped around because the ship would be moved but
not the shield.

This finalizes fixes to the boundary code so it moves a entities joint
entity along with it. It also applies a buffer to when the bounds set
the position so there's room for the joint when the position is set.

This was all done by adding the Joint component to both ship and shield
but only adding the shield joint to Box2d. Then the joint component is
added to Moveable. This means both ship and shield can access one another
so they can reset positions when bounds are hit.

I used a new function in Moveable that's called whenever setPos is
called, that only runs when the joint is set on Moveable. It uses a
offset of the entity to the joint which it uses when setting the new
position. This is to ensure the shield and entity are in the same
relative position to eachother once the position is reset.

Fixes #106.
Fixes #107.