Change horizontal direction when facing south
Has thrust check the `facingNorth()` function and passes it off
when getting force. This function will reverse the force x direction
by minus 1 when the control direction is top and down.

Had to change the oriented facing north function because the values
were not what I expected. I'm not sure what was going on here. I also
reversed facingNorth in thrust so it's techincally facing South. I did
this just because I think north facing should be what you use as a base.
Might make sense to change to facing south instead.

This just changes the x direction because I only want to modify that
direction when they're facing south. This means the player will have
their right and left thrusts switched when facing south, so it means
they're always the same physical direction, hopefully making it easier.

I also limited the facing north angle so it's not when the player is
facing west or east at all, otherwise the back and forward controls
are reversed which is way more confusing.