Add option in thrustable for strafing type
Adds option to have player use oriented strafing, which changes
strafing controls based on orientation. This will be used by
thrust to set actual controls.

This was done with a simple bool on the thrustable class for oriented
strafe which I defaulted to true because think it's the better option. I
put it there because all the configuration of oriented strafing is
already happening in thrust, not in controls. This also kinda works
because you also have to check orientation, which makes more sense
in thrust then in actual control code (probably).

I added a separate constructor for the bool property jsut for ease of
use at this point, and don't need option to set boolean after the
class is init.