Code to not render a hidden entity
I figured it'd be easiest to check if a component has a hidden component
check if that hidden component says to be hidden and then just stop
rendering.

This seems to be the fastest as it runs the least amount of code
rather then doing something at the render layer. It does require more
component checks, but not sure theres a way to get around that.