Arm system
Right now the system just binds weapon events and calls the activate
method on the armed component.

I couldn't get the final test working for `onAddEntity. It failed saying
that the `activate` method is not called. This is likely due to the
reference/casting involved in `getComponent`, I think that's what's
causing it to not work.

TODO: the armed component should only be called if the type of fired
event matches with the armed type of event.