Emit move event
When setting moveable so that the joint of an entity can listen when
it's parent moves, such as when bound resets the position.

I did this with the get user data and a static cast, but this should
likely be changed for performance reasons. I'm sure the user data and
static cast aren't the fastest method, so it might be better to pass the
reference to the entity in the setPos method, which kinda sucks from a
design perspective but is a better idea for perf.