Fix second entity bounds
It wasn't being reset on bounds because it didn't have an user data
because it wasn't damageable. Before I was only setting user data
on the box2d body to the current entity when it was a damageable
entity, so that in the collision handler in main, only damageable
entities would be paid attention to. But not I realize I want all
box2d moveable entities to have their user data set so you can always
get from a body to an entity like we need to do here.

The fix was to get rid of the set damageable method, as it was no
longer relevant to setting the user data and just do taht in the
moveable init method. This method now excepts an entity as it's last
param to allow setting this info.

TODO: I think I need to remove all box2d related stuff from moveable.