Load img
This uses SDL to load the image in the rectangle renderer. It uses
straight SDL code to do this there.

This loading should be moved to somewhere else so it can be mocked
and unit tested. It likely should be moved to the renderer, but if
it does, it should be moved there in a way that doesn't introduce
more SDL specific code.

Another problem arises with this design: what if we want a sprite
which is actually a polygon? It seems like something else should
be responsible for rendering sprites, not the shape code as they
only need to know the size of the sprite. This can maybe wait until
polygons are added.