Refactor rendering to remove SDL and such
Removes the rectangle renderer as it's more accurate to render based
on whether something is painted or sprited. This will then get the
bounding box from shaped to draw the actual sprite. This might have to be
changed if have to render the polygon as a bg image rather then rendering
everything as a sprite.

Refactors base renderer so it doesn't have any references to create an
SDL sprite. Not sure how this will work yet.