Fix shield blending
I changed it to add for now because it looks really cool and it works.
For some reason alpha bllending isn't working as expected anymore
with the sprite image and shield. I think the shield will have to be
a sprite at some point too so best to solve then.