Fix health not resetting to 100
Emit an impact event of 0 damage on death so the notification line
gets updated with the 100 health number immediately on death.

Before the notification wasn't getting updated because death event
was only adding a death to notification. So now just emit a fake
impact event after entity is reset so it will get new number.

Also tried to get impacts system to bind to death as well as impact
to update with newest health, but I think based on sequence of
events, the health isn't reset at that point yet (the impacts binding
is called before the death one).